+++ apps/web/src/app/(worker)/worker/admin-outlet/dashboard/page.tsx
@@ -16,6 +16,7 @@ import IconMenuDashboardMobile from "@/components/core/iconMenuDashboardMobile";
 import WeatherMobile from "@/components/core/weatherMobile";
 import WeatherWeb from "@/components/core/weatherWeb";
 import { useAdminOutletDashboardHook } from "@/features/adminOutlet/hooks/useAdminOutletDashboardHook";
+import HeaderAdminDashboardWeb from "@/components/core/headerDashboardWeb copy";
 
 
 export default function Page() {
@@ -86,31 +87,8 @@ export default function Page() {
 
             <main className="w-full h-full bg-neutral-200 p-4 gap-2 hidden md:flex flex-col">
                 <section className="w-full h-1/2 rounded-xl flex gap-2">
-                    <div className="w-full rounded-xl h-full flex items-center bg-orange-500 p-5">
-                        <div className="w-full h-fit">
-                            <div className="w-fit h-fit pb-5">
-                                <h1 className='font-bold border-b text-xl text-white pb-2'>Welcome, {name && name?.length > 10 ? name?.slice(0, 10) : name || 'Admin'}!</h1>
-                            </div>
-                            <div className="w-full">
-                                <p className="text-white">Pantau data order, dan kelola outlet satu tempat.</p>
-                                <p className="text-white pt-2">{isDayArr[isDay]} {isDate || '00/00/0000'}</p>
-                                <p className="text-white pt-2 flex gap-2 items-center"><span><FaStore className='text-lg' /></span> {storeName || 'CNC - Example'}</p>
-                            </div>
-                        </div>
-                        <div className="w-full h-full items-center flex justify-end">
-                            <Image
-                                className="h-[80%] w-fit"
-                                width={500}
-                                height={500}
-                                loading="lazy"
-                                alt="logo"
-                                src={'/images/charr.png'}
-                            />
-                        </div>
-                        <div className="flex items-start h-full">
-                            <NotificationOutletAdmin dataOrderNotif={dataOrderNotif} />
-                        </div>
-                    </div>
+
+                    <HeaderAdminDashboardWeb isDayArr={isDayArr} isDay={isDay} isDate={isDate} name={name} storeName={storeName} dataOrderNotif={dataOrderNotif} />
                     <div className="w-full rounded-xl h-full bg-gradient-to-tr from-sky-100 via-orange-100 to-white p-2 gap-2 flex items-center">
                         <WeatherWeb isCurrentWeather={isCurrentWeather} />
                         <div className="w-1/2 h-full bg-white bg-opacity-45 rounded-xl">
