+++ apps/web/src/components/core/MenuAdditionalDashboardMobile/type.ts
@@ -1,0 +1,10 @@
+interface IconItem {
+    icon: JSX.Element;  
+    url: string;        
+    name: string;       
+}
+
+
+export interface IconMenuDashboardMobileProps {
+    arrMenu: IconItem[];
+}
