+++ apps/web/src/components/core/headerDashboardWeb copy/type.ts
@@ -1,8 +1,9 @@
-interface HeaderDashboardWebProps {
+interface HeaderAdminDashboardWebProps {
     name: string;
     storeName: string;
     isDay: number;
     isDate: string;
     isDayArr: string[];
-    message:string
+    dataOrderNotif:any
 }
+
