+++ apps/web/src/features/superAdmin/components/EditOutletWeb/type.ts
@@ -1,0 +1,33 @@
+import { useRouter } from "next/navigation";
+
+export interface IEditOutletWeb {
+    isPosition: { lat: number; lng: number };
+        provinces: { province_id: string; province: string }[];
+        cities: { city_id: string; city_name: string }[];
+        provincesLoading: boolean;
+        citiesLoading: boolean;
+        dataOutlet: {
+            storeName: string;
+            address: string;
+            zipCode: string;
+            latitude: any;
+            longitude: any;
+        };
+        setSelectedProvince: React.Dispatch<React.SetStateAction<string>>;
+        isPositionCheck: boolean;
+        selectedProvince: string;
+        isDisabledSucces: boolean;
+        isPending: boolean;
+        router: ReturnType<typeof useRouter>;
+        handleUpdateOutlet: (values: {
+            storeName: string;
+            address: string;
+            province: string;
+            city: string;
+            zipCode: string;
+            latitude: string;
+            longitude: string;
+        }) => void;
+    }
+    
+
