+++ apps/web/src/features/superAdmin/components/ProductMobile/type.ts
@@ -1,0 +1,18 @@
+export interface IProductContent {
+    isPendingDelete: boolean;
+    handleUpdateItem: (product: any) => void;
+    prod: {
+        id: number;
+        itemName: string;
+        createdAt: string;
+    };
+    handleDeleteItem: (id: number) => void;
+    isPendingUpdate: boolean;
+    i:number
+}
+
+
+
+
+
+
